+++ jpf-api/src/main/java/com/joiest/jpf/entity/ShopCouponActiveInfo.java
@@ -63,7 +63,7 @@ public class ShopCouponActiveInfo {
     /**
      * 豆数量
      */
-    private Integer dou;
+    private BigDecimal dou;
 
     /**
      * 根据行为不同显示的内容不同
@@ -201,11 +201,11 @@ public class ShopCouponActiveInfo {
         this.money = money;
     }
 
-    public Integer getDou() {
+    public BigDecimal getDou() {
         return dou;
     }
 
-    public void setDou(Integer dou) {
+    public void setDou(BigDecimal dou) {
         this.dou = dou;
     }
